Changes to XML and command definition to bring the tool more in-line with user interface expectations in GalaxyTrakr.
More use of Cheetah templating where appropriate so that variable substitutions are clearer in the job's recorded command line invocation (useful for debugging.)

Reference selection interface consistent with other tools; allows the selection of either a FASTA reference file from the user's own history or from a centrally-curated table of GalaxyTrakr reference assemblies.

Tool now allows for the selection of a list of assemblies from the history or a list of paired-end reads; lists-of-pairs are our preferred format for collections of accessioned sequencing data and lists of assemblies are generally the output when assemblers are run on such collections.

Use of && rather than ; in the Bash command template allows for preservation of the exit code.

It's not necessary to pipe STDOUT to a dataset item; Galaxy jobs capture and record both STDOUT and STDERR streams. At the end of the pipeline, we now cat the nextflow log into STDOUT for better capture.

Optional pipeline tweaking parameters are now enclosed in an optional section to reduce UI clutter.

Test cases now include assembly and reads cases in separate tests. Generally we consider these options to be modal in GalaxyTrakr; users don't generally have heterogenous collections of reads and assemblies.

The conventional data format for FASTQ sequencing data in Galaxy, for various legacy reasons, is "fastqsanger".

Can verify tests pass (more or less, since the order of the Screening Results file changes every time.)

bugfixes